# Propagate names through `r_lgl_which()`