--- a/NEWS.md
+++ b/NEWS.md
@@ -1,8 +1,10 @@
 # rlang (development version)
 
+* The C API function `r_lgl_which()` now propagates the names of the input
+  (#1471).
+
 * The `pkg_version_info()` function now allows `==` for package
   version comparison (#1469, @kryekuzhinieri).
-
 
 # rlang 1.0.5