- Cleaned up the style in `r_lgl_which()`
- Lifted the `na_propagate` branch out of the for loop
- Simplified the if condition, especially in the `na_propagate = FALSE` case
- Added in support for propagating names

I don't think this has much performance impact when there aren't names. In fact, it looks like moving the `na_propagate` branch out of the loop and simplifying the if condition has made things slightly faster in the `na_propagate = FALSE` case.

``` r
devtools::load_all()
#> ℹ Loading rlang

# Uniform mix
x <- sample(c(TRUE, FALSE, NA), size = 2e6, replace = TRUE)

bench::mark(
  r_lgl_which(x, na_propagate = TRUE),
  r_lgl_which(x, na_propagate = FALSE),
  iterations = 500,
  check = FALSE
)

# Main
#> # A tibble: 2 × 6
#>   expression                                min   median
#>   <bch:expr>                           <bch:tm> <bch:tm>
#> 1 r_lgl_which(x, na_propagate = TRUE)   10.34ms   12.5ms
#> 2 r_lgl_which(x, na_propagate = FALSE)   9.44ms   10.5ms
#> # … with 3 more variables: `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>

# This PR
#> # A tibble: 2 × 6
#>   expression                                min   median
#>   <bch:expr>                           <bch:tm> <bch:tm>
#> 1 r_lgl_which(x, na_propagate = TRUE)    9.67ms  12.02ms
#> 2 r_lgl_which(x, na_propagate = FALSE)   6.13ms   6.71ms
#> # … with 3 more variables: `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>

# Mostly `TRUE`
x <- sample(c(rep(TRUE, 100), FALSE, NA), size = 2e6, replace = TRUE)

bench::mark(
  r_lgl_which(x, na_propagate = TRUE),
  r_lgl_which(x, na_propagate = FALSE),
  iterations = 500,
  check = FALSE
)

# Main
#> # A tibble: 2 × 6
#>   expression                                min   median
#>   <bch:expr>                           <bch:tm> <bch:tm>
#> 1 r_lgl_which(x, na_propagate = TRUE)    3.88ms   4.82ms
#> 2 r_lgl_which(x, na_propagate = FALSE)   3.34ms   3.98ms
#> # … with 3 more variables: `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>

# This PR
#> # A tibble: 2 × 6
#>   expression                                min   median
#>   <bch:expr>                           <bch:tm> <bch:tm>
#> 1 r_lgl_which(x, na_propagate = TRUE)    4.16ms   4.83ms
#> 2 r_lgl_which(x, na_propagate = FALSE)   2.75ms   3.31ms
#> # … with 3 more variables: `itr/sec` <dbl>,
#> #   mem_alloc <bch:byt>, `gc/sec` <dbl>
```